Removed some foreign key constraints and returned some primary keys to integer (#8)
* Added initial files structure

* Modified env.py to deal with hypertables

* Modified index creation in env.py

* Cleaned the files and added comments

* Added documentation files

* Added license file

* Added more documentation files

* Added .env file, removed root path and modified the documentation files

* Removed .env file

* Added database models

* Added more database models

* Completed hegemony cone endpoint

* Fixed foreign keys errors in models

* Added missing indexes to models

* Untracked migration files

* Renamed to hegemony_cone.py and added alembic/versions to gitignore

* Returned some id(primary key) of tables back from BigInteger to Integer

* Removed foreign key constraints from some models

+++ models/disco_probes.py
@@ -12,7 +12,7 @@ class DiscoProbes(Base):
         'columns': ['event_id']
     }]
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
 
     probe_id = Column(Integer, nullable=False,
                       doc='Atlas probe ID of disconnected probe.')
@@ -38,7 +38,7 @@ class DiscoProbes(Base):
     lon = Column(Float, default=0.0, nullable=False,
                  doc='Longitude of the probe during the network detection as reported by RIPE Atlas.')
 
-    event_id = Column(BigInteger,
+    event_id = Column(Integer,
                       ForeignKey('ihr_disco_events.id', ondelete='CASCADE',
                                  name='fk_disco_probes_event_id'),
                       nullable=False,
