+++ models/tr_hegemony_identifier.py
@@ -5,7 +5,7 @@ from config.database import Base
 class TRHegemonyIdentifier(Base):
     __tablename__ = 'ihr_tr_hegemony_identifier'
 
-    id = Column(BigInteger, autoincrement=True, primary_key=True)
+    id = Column(Integer, autoincrement=True, primary_key=True)
 
     name = Column(String(255), nullable=False,
                   doc='Value of the identifier. The meaning depends on the identifier type: <ul><li>type=AS: ASN</li><li>type=IX: PeeringDB IX ID</li><li>type=MB: IXP member (format: ix_id;asn)</li><li>type=IP: Interface IP of an IXP member</li></ul>')
